refactor: rename `server` to `core`
