+++ src/core/errors.ts
