+++ src/core/lib/csrf-token.ts
