+++ src/core/lib/extend-res.ts
