+++ src/core/lib/oauth/client-legacy.ts
