+++ src/core/pages/index.ts
