--- src/server/index.ts
@@ -1,2 +1,0 @@
-export type { IncomingRequest, OutgoingResponse } from "./handler"
-export { default } from "../next"
